People as Element and settings

+++ src/migrations/m250911_132206_people_element_2.php
@@ -25,8 +25,10 @@ class m250911_132206_people_element_2 extends Migration
            -> delete('{{%museumplus_people}}', 'id < :id', [':id' => $firstPerson->id])
             ->execute();
 
-        $this->dropColumn('{{%museumplus_people}}', 'title');
-
+        $schema = Craft::$app->getDb()->getSchema()->getTableSchema('{{%museum_people}}');
+        if($schema->getColumn('title')) {
+            $this->dropColumn('{{%museumplus_people}}', 'title');
+        }
 
         $this->addForeignKey(
                $this->db->getForeignKeyName(),
